Added flush capability.
Required by: BZ 647331 -- memory structures not flushed on sysrq.

This adds rhts-flush to be used with rhts tests.

Why?

I do not want to make them dependent on beah, as it's better to define
a hook to be redefined by anyone using rhts tests.

If the hook is not defined this will result in message in stderr which
may be beneficial.