runners: initialize supplementary groups
Normally this happens in /bin/login. Beah tasks are running
via init scripts, and this leaves supplementary groups empty.
This is causing problems, because some ltp tests may unexpectedly
fail when ran through beah, but succeed when ran directly from
shell. For example: chown03 from ltp 20110228

To fix this, check if list of groups is empty, and in that case
initialize it according to current euid.

Change-Id: Idc97fcd3050af3aece4ade9a86664d953ef6f55c
Signed-off-by: Jan Stancek <jstancek@redhat.com>
Bug: 706026