Make Beah work in an IPv6 only environment
This patch allows Beah to function in an IPv6 only environment. The
exact use case in which this has been tested is the test system
starts in a dual IPv4/IPv6 environment and IPv4 is disabled so that
a test runs in an IPv6 only environment. The test runner can then
return the system and the logs should be submited - the usual tasks
that Beah does. So, all in all, it should work as it does with
IPv4.

Bug: 810893