Implement a new config option to use IPv4 only
This patch implements the support for a new config option: IPV6_DISABLED
for beah_beaker.conf and beah.conf. If this option is set to True, Beah
will not use IPv6 for any network communication, even if otherwise
possible. By default, it is set to False.

Beaker side of this fix added support for a new ``ks_meta`` option:
``beah_no_ipv6`` to add this configuration to Beah.

Bug: 1065811
Change-Id: I29de0371910ebf62b349f9a3a5693bd0e65b00d1