Waiting and Running are not the only states for an incomplete recipe
Specifically, Installing will soon be a thing (bug 991245) although if
the kickstart %post works properly the recipe should be back to Waiting
by the time beah sees it.

In any case, we can flip this conditional so that it looks for statuses
indicating that the recipe is actually finished, and otherwise assumes
it's not.

This particular check is not strictly necessary anyway -- the status of
each individual task is also checked to ensure that they are never
re-run. The only purpose this check serves is to give a slightly cleaner
message in beah's logs in case it's started up after the recipe is
finished (e.g. powering up a released system). So it's okay if we aren't
too strict with this conditional.

Change-Id: Ifb9e12443d7c608a5008bfbd5547a5a7fa8bf362